pyproject.toml
[project]
name = "tool-calling-llms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain[google-genai,openai]>=1.0.1",
    "python-dotenv>=1.1.1",
    "rich>=14.2.0",
]
